# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/api/game/start": {
            "post": {
                "description": "받은 명세서에 알맞게 게임 시작",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "게임 시작",
                "parameters": [
                    {
                        "description": "게임 명세",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.GameStartReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.GameInfo": {
            "type": "object",
            "properties": {
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.PlayerInfo"
                    }
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "main.GameStartReq": {
            "type": "object",
            "properties": {
                "game_info": {
                    "$ref": "#/definitions/main.GameInfo"
                },
                "map_info": {
                    "$ref": "#/definitions/main.MapInfo"
                },
                "timeout": {
                    "type": "integer"
                }
            }
        },
        "main.MapInfo": {
            "type": "object",
            "properties": {
                "floor": {
                    "type": "integer"
                },
                "horizontal": {
                    "type": "integer"
                },
                "locktime": {
                    "type": "integer"
                },
                "vertical": {
                    "type": "integer"
                }
            }
        },
        "main.PlayerInfo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "stat": {
                    "$ref": "#/definitions/main.Stat"
                }
            }
        },
        "main.Stat": {
            "type": "object",
            "properties": {
                "int": {
                    "type": "integer"
                },
                "pow": {
                    "type": "integer"
                },
                "spd": {
                    "type": "integer"
                }
            }
        }
    }
}